requirements.txt
pytorch==2.0.0
pandas==1.5.3
numpy==1.23.5
tqdm==4.65.0
scikit-learn==1.2.2